Added more tests around giving users roles, moved policies and schemes to new namespaces
