+++ tests/Integration/CreatesUser.php
@@ -22,13 +22,15 @@ trait CreatesUser
      * Creates a user for use in integration tests
      *
      * @param string $password The user's password
+     * @param string|list<string> $roles The user's roles
      * @return User The created user
      * @throws FailedContentNegotiationException|SerializationException|HttpException|RuntimeException|Exception Thrown if there was an error creating the user
      */
-    private function createUser(string $password = 'password'): User
+    private function createUser(string $password = 'password', string|array $roles = []): User
     {
         // Create a unique email address so we do not have collisions
-        $newUser = new NewUser(\bin2hex(\random_bytes(8)) . '@example.com', $password);
+        $roles = \is_string($roles) ? [$roles] : $roles;
+        $newUser = new NewUser(\bin2hex(\random_bytes(8)) . '@example.com', $password, $roles);
         $newUserResponse = $this->post('/users', body: $newUser);
 
         if ($newUserResponse->getStatusCode() !== HttpStatusCode::Ok) {
